Added First Neighborhood.
Along with adding an intro scene. I added the entire first neighborhood scene. The scene consists of some houses, your car, your house, and your mailbox. Hovering over every item in the scene provides a description, and some descriptions change depending on your items. The goal of this scene is to give your mom the mail in the mail box, and drive to the next scene.