Fleshed out a new scene and created 2 more.
I created a new scene that lets you travel to two new scenes (Road, Hills, and Farm). This interconnected scene has been fleshed out with houses, a road, and visual icons to represent the two linked scenes. It also links back to the first scene (Home). The Hills and Farm scene were created, but nothing is in them except for a background pallete.